# scripts/vim-patch.sh: improve performance