@blueyed 2019-06-07
> Why WIP?

Different output with regard to the mentioned patches (commits with duplicate tags).

One possibility I see would be using not `refs/tags` in the `for-each-ref` for example, but wanted to wait how the duplicate tags get handled in Vim.