@blueyed 2019-06-07
Also I've thought about setting up some kind of service / regularly generated static page with e.g. links etc from there, where it could use an additional (SQLite) DB to store some state / more information then.