@justinmk 2019-06-07
> Different output with regard to the mentioned patches (commits with duplicate tags).

If Vim starts to do it a lot then we might need a different approach (e.g. ignore tags and only use commit-ids). Otherwise there's nothing we can do.

> Also I've thought about setting up some kind of service / regularly generated static page with e.g. links etc from there, where it could use an additional (SQLite) DB to store some state / more information then.

https://neovim.io/doc/reports/vimpatch/ could use some attention :)

sqlite is out of scope for these reports, we don't want to maintain servers.