@blueyed 2019-06-08
Sorting refs by authordate produces the same result/output now.