@blueyed 2019-06-08
> https://neovim.io/doc/reports/vimpatch/ could use some attention :)
> It is generated by https://github.com/neovim/bot-ci/blob/master/ci/vimpatch-report.sh

How/when is it being run / triggered?

> sqlite is out of scope for these reports, we don't want to maintain servers.

State could always be stored e.g. in a Git repo or something similar also.
I could imagine that flagging / tagging patches might be useful, but that could also be done automatically.  E.g. it might be interesting to check if a patch applies cleanly already, and then maybe even create automated PRs based on that.