@justinmk 2019-06-08
> How/when is it being run / triggered?

It is run daily at https://github.com/neovim/bot-ci . The README there gives an overview; the code is just a flat list of scripts, https://github.com/neovim/bot-ci/tree/master/ci  (one script for each job-type).

It is triggered by an AWS lambda task, which is simply [this script](https://github.com/neovim/bot-ci/blob/master/scripts/trigger-build.py) copy-pasted into AWS .

> State could always be stored e.g. in a Git repo or something similar also.

We do have some static state in the https://github.com/neovim/doc/ repo. But it is better to store "flags" and other knowledge (state) in the VCS logs of `neovim` itself, as we already do with `vim-patch:xxx` tokens. The core `neovim` repo should be the source of truth (wherever that rule isn't followed is just a "todo" from the early days ;).

> I could imagine that flagging / tagging patches might be useful, but that could also be done automatically. E.g. it might be interesting to check if a patch applies cleanly already, and then maybe even create automated PRs based on that.

Sure, for auto-creating PRs https://github.com/neovim/bot-ci/blob/master/ci/auto-pullrequest.sh shows an example. We want to do auto-PRs  for more purposes (e.g. updating Unicode definitions).