@blueyed 2019-06-08
Thanks for all the info!

> I'm not really sure it will help much to auto-apply vim patches, though. I would prefer to spend time elsewhere.

By applying them manually? ;D

(I get your point though.. but think that eventually we'll be in a state where there are not >1k patches missing, and patch adjustment works so good that new Vim patches could be auto-applied on a daily basis)