@justinmk 2019-06-08
> eventually we'll be in a state where there are not >1k patches missing, and patch adjustment works so good that new Vim patches could be auto-applied on a daily basis)

Sure. Everything would be easier if we weren't behind, so help with this is really appreciated. We've made a ton of progress thanks to @janlazo and others. The automation helps too, even what little there is.